+++ src/platform_impl/web/web_sys/scaling.rs
@@ -70,8 +70,7 @@ impl ScaleChangeDetectorInternal {
         mql
     }
 
-    fn handler(&mut self, event: MediaQueryListEvent) {
-        assert!(!event.matches());
+    fn handler(&mut self, _event: MediaQueryListEvent) {
         let mql = self
             .mql
             .take()